rename debug into kuzzleDebug
