+++ lib/cluster/masterNode.js
@@ -1,7 +1,7 @@
 'use strict';
 
 let
-  debug = require('../debug')('kuzzle:cluster:node:master'),
+  debug = require('../kuzzleDebug')('kuzzle:cluster:node:master'),
   Promise = require('bluebird'),
   Node = require('./node');
