+++ lib/cluster/node.js
@@ -1,7 +1,7 @@
 'use strict';
 
 let
-  debug = require('../debug')('kuzzle:cluster:node'),
+  debug = require('../kuzzleDebug')('kuzzle:cluster:node'),
   debugMerge = require('debug')('kuzzle:cluster:merge'),
   Request = require('kuzzle-common-objects').Request,
   RequestContext = require('kuzzle-common-objects').models.RequestContext;
