+++ lib/cluster/slaveNode.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const
-  debug = require('../debug')('kuzzle:cluster:node:slave'),
+  debug = require('../kuzzleDebug')('kuzzle:cluster:node:slave'),
   Node = require('./node');
 
 /**
