+++ lib/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 let
-  debug = require('./debug')('kuzzle:cluster'),
+  debug = require('./kuzzleDebug')('kuzzle:cluster'),
   os = require('os'),
   InternalError = require('kuzzle-common-objects').errors.InternalError,
   MasterNode = require('./cluster/masterNode'),
